# ⚡️ hits resolution performance boost