--- a/modules/server/src/download/index.js
+++ b/modules/server/src/download/index.js
@@ -60,7 +60,6 @@ export default function({ projectId, io }) {
   router.use(bodyParser.urlencoded({ extended: true }));
 
   router.post('/', async function(req, res) {
-    console.time('download');
     const { params, downloadKey } = req.body;
     const { files, fileName = 'file.tar.gz', mock, chunkSize } = JSON.parse(
       params,
@@ -89,7 +88,6 @@ export default function({ projectId, io }) {
         `attachment; filename=${responseFileName}`,
       );
       output.pipe(res).on('finish', () => {
-        console.timeEnd('download');
         io.emit(`server::download::${downloadKey}`);
       });
     }