--- a/modules/server/src/utils/config.js
+++ b/modules/server/src/utils/config.js
@@ -4,3 +4,5 @@ export const PROJECT_ID = process.env.PROJECT_ID;
 export const PING_MS = process.env.PING_MS || 2200;
 export const ES_LOG = process.env.ES_LOG?.split?.(',') || 'error';
 export const MAX_LIVE_VERSIONS = process.env.MAX_LIVE_VERSIONS || 3;
+export const DOWNLOAD_STREAM_BUFFER_SIZE =
+  process.env.DOWNLOAD_STREAM_BUFFER_SIZE || 2000;