--- a/modules/server/src/utils/dataToTSV.js
+++ b/modules/server/src/utils/dataToTSV.js
@@ -106,6 +106,8 @@ export default function({ columns, index, uniqueBy, emptyValue = '--' }) {
   let isFirst = true;
 
   return through2.obj(function(data, enc, callback) {
+    const timer = `dataToTSV_${Date.now()}`;
+    console.time(timer);
     if (isFirst) {
       isFirst = false;
       this.push(columnsToHeader({ columns }));
@@ -115,6 +117,7 @@ export default function({ columns, index, uniqueBy, emptyValue = '--' }) {
 
     if (rows) {
       this.push(rows);
+      console.timeEnd(timer);
     }
 
     callback();